backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
geoalchemy2==0.14.7
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic>=2.9.0,<3.0.0
pydantic-settings==2.2.1
pydantic[email]>=2.9.0,<3.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.1
aioredis==2.0.1
aiosmtplib==3.0.1
email-validator==2.1.0.post1
twilio==8.10.0
boto3==1.34.32
s3fs==2024.3.1
